rename update.py to not clash with update function, update ci tests
