+++ .github/test_conda_env.yml
@@ -4,5 +4,4 @@ channels:
 dependencies:
   - numpy
   - pandas
-  - requests
   - fortran-compiler
