+++ pygtide/__init__.py
@@ -2,5 +2,5 @@ from pygtide.core import pygtide
 from pygtide.core import predict_series, predict_spectrum, predict_table
 from pygtide.core import plot_series, plot_spectrum
 from pygtide.tests import test
-from pygtide.update import update_etpred_db, update
+from pygtide.update_etpred_data import update
 __version__ = '0.6.0'
